more progress here
