--- internal/cmd/destroy.go
@@ -1,24 +1,0 @@
-package cmd
-
-import (
-	"github.com/gobuffalo/cli/internal/cmd/destroy"
-	"github.com/spf13/cobra"
-)
-
-// DestroyCmd destroys generated code
-var DestroyCmd = &cobra.Command{
-	Use:     "destroy",
-	Short:   "Destroy generated components",
-	Aliases: []string{"d"},
-}
-
-func init() {
-	DestroyCmd.AddCommand(destroy.ResourceCmd)
-	DestroyCmd.AddCommand(destroy.ActionCmd)
-	DestroyCmd.AddCommand(destroy.MailerCmd)
-
-	DestroyCmd.PersistentFlags().BoolVarP(&destroy.YesToAll, "yes", "y", false, "confirms all beforehand")
-
-	decorate("destroy", DestroyCmd)
-	RootCmd.AddCommand(DestroyCmd)
-}
