--- internal/cmd/generate.go
@@ -1,22 +1,0 @@
-package cmd
-
-import (
-	"github.com/gobuffalo/cli/internal/cmd/generate"
-	"github.com/spf13/cobra"
-)
-
-var generateCmd = &cobra.Command{
-	Use:     "generate",
-	Short:   "Generate application components",
-	Aliases: []string{"g"},
-}
-
-func init() {
-	generateCmd.AddCommand(generate.ResourceCmd)
-	generateCmd.AddCommand(generate.ActionCmd)
-	generateCmd.AddCommand(generate.TaskCmd)
-	generateCmd.AddCommand(generate.MailCmd)
-	decorate("generate", generateCmd)
-
-	RootCmd.AddCommand(generateCmd)
-}
