+++ internal/cmd/popinstructions.txt
@@ -1,0 +1,8 @@
+Pop support has been moved to the https://github.com/gobuffalo/buffalo-pop plugin.
+
+!! PLEASE READ PLUGIN DOCUMENTATION - https://gobuffalo.io/en/docs/plugins
+
+Buffalo Plugins Installation*:
+
+	$ buffalo plugins install github.com/gobuffalo/buffalo-pop
+
