+++ internal/cmd/root.go
@@ -1,16 +1,37 @@
 package cmd
 
 import (
+	_ "embed"
 	"fmt"
 	"os"
 	"strings"
 
+	"github.com/gobuffalo/cli/internal/cmd/build"
+	"github.com/gobuffalo/cli/internal/cmd/destroy"
+	"github.com/gobuffalo/cli/internal/cmd/fix"
+	"github.com/gobuffalo/cli/internal/cmd/generate"
+	"github.com/gobuffalo/cli/internal/cmd/info"
+	"github.com/gobuffalo/cli/internal/cmd/new"
+	cmdplugins "github.com/gobuffalo/cli/internal/cmd/plugins"
+	"github.com/gobuffalo/cli/internal/cmd/setup"
+	"github.com/gobuffalo/cli/internal/cmd/test"
+	"github.com/gobuffalo/cli/internal/cmd/version"
 	"github.com/gobuffalo/cli/internal/plugins"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 )
 
-var anywhereCommands = []string{"new", "version", "info", "help"}
+const (
+	dbNotFound  = `unknown command "db"`
+	popNotFound = `unknown command "pop"`
+)
+
+var (
+	anywhereCommands = []string{"new", "version", "info", "help"}
+
+	//go:embed popinstructions.txt
+	popInstallInstructions string
+)
 
 // RootCmd is the hook for all of the other commands in the buffalo binary.
 var RootCmd = &cobra.Command{
@@ -44,6 +65,36 @@ var RootCmd = &cobra.Command{
 // Execute adds all child commands to the root command sets flags appropriately.
 // This is called by main.main(). It only needs to happen once to the rootCmd.
 func Execute() {
+	newCmd := new.Cmd()
+	setupCmd := setup.Cmd()
+	generateCmd := generate.Cmd()
+	destroyCmd := destroy.Cmd()
+	versionCmd := version.Cmd()
+	testCmd := test.Cmd()
+
+	decorate("new", newCmd)
+	decorate("info", RootCmd)
+	decorate("fix", RootCmd)
+	decorate("update", RootCmd)
+	decorate("setup", setupCmd)
+	decorate("generate", generateCmd)
+	decorate("destroy", destroyCmd)
+	decorate("version", versionCmd)
+	decorate("test", testCmd)
+
+	RootCmd.AddCommand(newCmd)
+	RootCmd.AddCommand(build.Cmd())
+	RootCmd.AddCommand(cmdplugins.PluginsCmd)
+	RootCmd.AddCommand(info.Cmd())
+	RootCmd.AddCommand(setupCmd)
+	RootCmd.AddCommand(generateCmd)
+	RootCmd.AddCommand(fix.Cmd())
+	RootCmd.AddCommand(destroyCmd)
+	RootCmd.AddCommand(versionCmd)
+	RootCmd.AddCommand(testCmd)
+
+	decorate("root", RootCmd)
+
 	if err := RootCmd.Execute(); err != nil {
 		if strings.Contains(err.Error(), dbNotFound) || strings.Contains(err.Error(), popNotFound) {
 			logrus.Errorf(popInstallInstructions)
@@ -56,22 +107,6 @@ func Execute() {
 		}
 		os.Exit(-1)
 	}
-}
-
-const dbNotFound = `unknown command "db"`
-const popNotFound = `unknown command "pop"`
-const popInstallInstructions = `Pop support has been moved to the https://github.com/gobuffalo/buffalo-pop plugin.
-
-!! PLEASE READ PLUGIN DOCUMENTATION - https://gobuffalo.io/en/docs/plugins
-
-Buffalo Plugins Installation*:
-
-	$ buffalo plugins install github.com/gobuffalo/buffalo-pop
-
-`
-
-func init() {
-	decorate("root", RootCmd)
 }
 
 func insideBuffaloProject() bool {
