+++ internal/cmd/test/cmd.go
@@ -1,0 +1,7 @@
+package test
+
+import "github.com/spf13/cobra"
+
+func Cmd() *cobra.Command {
+	return cmd
+}
