+++ internal/cmd/test/test_test.go
@@ -1,4 +1,4 @@
-package cmd
+package test
 
 import (
 	"reflect"
